Fractional cell counts beside the view
1.5 x 1 x 1 is half a cell more of a framework, which is how you see
what a pore connects to without the eight-fold picture a whole extra
cell in every direction gives.  The Display range dialog could always
say 1.5; the boxes on the toolbar could not.

They are double spinboxes now, set_cells takes floats, and the cell
outline is clipped to the range rather than rounded up to whole boxes
-- a whole box drawn round 1.5 cells claims a cell of crystal that is
not on screen.  Keyboard tracking is off, because typing "1.5" passes
through 1 and 15 and each would rebuild the scene and reframe.

+++ xtalapp/document.py
@@ -1794,7 +1794,7 @@ class Document(QObject):
             setattr(self.view, key, value)
         self.viewChanged.emit()
 
-    def set_cells(self, na: int, nb: int, nc: int) -> None:
+    def set_cells(self, na: float, nb: float, nc: float) -> None:
         self.view.set_cells(na, nb, nc)
         self.viewChanged.emit()
